gallerie a revoir car sauvegarde pas ok. images referencees dans la bdd

+++ app/src/main/java/com/example/mycomics/fragments/PictureFragment.java
@@ -3,6 +3,8 @@ package com.example.mycomics.fragments;
 import static androidx.navigation.fragment.FragmentKt.findNavController;
 
 import android.content.ContentValues;
+import android.content.Context;
+import android.content.Intent;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.graphics.Matrix;
@@ -30,12 +32,18 @@ import android.widget.Toast;
 
 import com.example.mycomics.R;
 import com.example.mycomics.databinding.FragmentPictureBinding;
+import com.example.mycomics.helpers.DataBaseHelper;
 import com.google.common.util.concurrent.ListenableFuture;
 
+import java.io.ByteArrayOutputStream;
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executor;
 
@@ -50,11 +58,18 @@ public class PictureFragment extends Fragment {
 
 
     //* ----------------------------------------------------------------------------------------- */
+    //* Database handler initialization
+    //* ----------------------------------------------------------------------------------------- */
+    DataBaseHelper dataBaseHelper;
+
+
+    //* ----------------------------------------------------------------------------------------- */
     //* Camera Provider declarations
     //* ----------------------------------------------------------------------------------------- */
     private ListenableFuture<ProcessCameraProvider> cameraProviderFuture;
     private ImageCapture imageCapture;
     private String picturePath = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES) + "/MyComics/";
+    private String imgName = "";
 //    private File imgFile = null;
 
 
@@ -114,6 +129,7 @@ public class PictureFragment extends Fragment {
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
+        dataBaseHelper = new DataBaseHelper(getContext());
 
         /* Data sent from source fragments */
         Integer book_id = getArguments().getInt("book_id");
@@ -122,10 +138,12 @@ public class PictureFragment extends Fragment {
         binding.btnPictureAddPicture.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+                // Delete last temporary taken picture if multiple shots
                 File imgFileOld = new File(picturePath + "temp.jpg");
                 if (imgFileOld.exists()) {
                     imgFileOld.delete();
                 }
+                // Capture new picture
                 capturePicture();
             }
         });
@@ -133,19 +151,75 @@ public class PictureFragment extends Fragment {
         binding.btnPictureSave.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+                // Fetch last temporary picture
                 File imgToSave = new File(picturePath + "temp.jpg");
+                // If there is a taken picture
                 if (imgToSave.exists()){
-                    // Delete old picture
-                    File imgFileOld = new File(picturePath + "book_" + book_id + ".jpg");
-                    if (imgFileOld.exists()) {
-                        imgFileOld.delete();
+                    // Get old book picture
+                    String imgToDelete = null;
+                    try {
+                        imgToDelete = dataBaseHelper.getBookById(book_id).getBook_picture();
+                    } catch (Exception e) {
+                        // no picture in database for this book
                     }
-
-                    File newFile = new File(picturePath + "book_" + book_id + ".jpg");
-                    imgToSave.renameTo(newFile);
-
+                    if (imgToSave.length() > 0) { // the database has a picture referenced
+                        File imgFileOld = new File(picturePath + imgToDelete);
+                        if (imgFileOld.exists()) { // The picture physically exists
+                            // delete old physical image file
+                            imgFileOld.delete();
+                        }
+                    }
                 }
-
+                    // we'll start with the temporary picture already open to a file
+                    Bitmap imgOrigin = BitmapFactory.decodeFile(imgToSave.getAbsolutePath());
+                    // original sizes
+                    int originWidth = imgOrigin.getWidth();
+                    int originHeight = imgOrigin.getHeight();
+                    // size wanted
+                    final int newWidth = 600;
+                    Bitmap imgNew;
+                    if(originWidth > newWidth) {
+                        // reduce size if needed
+                        int newHeight = originHeight / (originWidth / newWidth);
+                        imgNew = Bitmap.createScaledBitmap(imgOrigin, newWidth, newHeight, false);
+                    } else { // size stays the same
+                        imgNew = Bitmap.createScaledBitmap(imgOrigin, originWidth, originHeight, false);
+                    }
+                    ByteArrayOutputStream outStream = new ByteArrayOutputStream();
+                    imgNew.compress(Bitmap.CompressFormat.JPEG,70 , outStream);
+                    String timeStamp=new SimpleDateFormat("yyyyMMdd_HHmmss").format(new Date());
+                    imgName = "BookId_" + book_id + "_" + timeStamp + ".jpg";
+                    File imgResized = new File(picturePath + imgName);
+                    try {
+                        imgResized.createNewFile();
+                    } catch (IOException e) {
+                        throw new RuntimeException(e);
+                    }
+                    //write the bytes in file
+                    FileOutputStream fo = null;
+                    try {
+                        fo = new FileOutputStream(imgResized);
+                    } catch (FileNotFoundException e) {
+                        throw new RuntimeException(e);
+                    }
+                    try {
+                        fo.write(outStream.toByteArray());
+                    } catch (IOException e) {
+                        throw new RuntimeException(e);
+                    }
+                    // Close de FileOutput
+                    try {
+                        fo.close();
+                    } catch (IOException e) {
+                        throw new RuntimeException(e);
+                    }
+                    // Delete origin file
+                    imgToSave.delete();
+                    // update book picture in database
+                    dataBaseHelper.updateBookPicture(dataBaseHelper,book_id,  imgName);
+                    // update gallery
+                    //sendBroadcast(new Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, Uri.fromFile(myNewFile)));
+                    refreshGallery(getContext(), imgResized);
                 // Data bundle storing search string
                 Bundle bundle = new Bundle();
                 bundle.putInt("book_id", book_id);
@@ -178,6 +252,17 @@ public class PictureFragment extends Fragment {
                     bmp.compress(Bitmap.CompressFormat.JPEG,100,os);
                 }
                 pictureRefreshScreen();
+            }
+        });
+
+        binding.btnPictureAbort.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                // Data bundle storing search string
+                Bundle bundle = new Bundle();
+                bundle.putInt("book_id", book_id);
+                // Go to SearchResultFragment with the data bundle
+                findNavController(PictureFragment.this).navigate(R.id.action_picture_to_bookDetail, bundle);
             }
         });
 
@@ -265,7 +350,10 @@ public class PictureFragment extends Fragment {
                 }
         );
     }
-
-
-
+    public static void refreshGallery(Context mContext, File file) {
+        Intent mediaScanIntent = new Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE);
+        Uri contentUri = Uri.fromFile(file);
+        mediaScanIntent.setData(contentUri);
+        mContext.sendBroadcast(mediaScanIntent);
+    }
 }
